# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProofDesk API - AI-powered contract and document review SaaS
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

export const DocumentStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export type DocumentOverallRisk =
  | (typeof DocumentOverallRisk)[keyof typeof DocumentOverallRisk]
  | null;

export const DocumentOverallRisk = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Document {
  id: number;
  title: string;
  documentType: string;
  status: DocumentStatus;
  content?: string | null;
  overallRisk?: DocumentOverallRisk;
  summary?: string | null;
  findingCount: number;
  criticalCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDocumentBody {
  title: string;
  documentType: string;
  content: string;
}

export interface DocumentList {
  documents: Document[];
  total: number;
}

export type FindingSeverity =
  (typeof FindingSeverity)[keyof typeof FindingSeverity];

export const FindingSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Finding {
  id: number;
  documentId: number;
  documentTitle?: string | null;
  category: string;
  severity: FindingSeverity;
  title: string;
  description: string;
  recommendation: string;
  clause?: string | null;
  createdAt: string;
}

export interface FindingList {
  findings: Finding[];
  total: number;
}

export type DocumentReviewOverallRisk =
  | (typeof DocumentReviewOverallRisk)[keyof typeof DocumentReviewOverallRisk]
  | null;

export const DocumentReviewOverallRisk = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface DocumentReview {
  document: Document;
  findings: Finding[];
  summary?: string | null;
  overallRisk?: DocumentReviewOverallRisk;
}

export interface DashboardSummary {
  totalDocuments: number;
  completedReviews: number;
  pendingReviews: number;
  totalFindings: number;
  criticalFindings: number;
  highRiskDocuments: number;
  avgFindingsPerDoc: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  uploaded: "uploaded",
  analyzed: "analyzed",
  finding_added: "finding_added",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  documentId: number;
  documentTitle: string;
  description: string;
  severity?: string | null;
  createdAt: string;
}

export interface ActivityList {
  activities: ActivityItem[];
}

export type RiskBreakdownBySeverityItem = {
  severity: string;
  count: number;
};

export type RiskBreakdownByCategoryItem = {
  category: string;
  count: number;
};

export type RiskBreakdownByDocumentTypeItem = {
  documentType: string;
  count: number;
};

export interface RiskBreakdown {
  bySeverity: RiskBreakdownBySeverityItem[];
  byCategory: RiskBreakdownByCategoryItem[];
  byDocumentType: RiskBreakdownByDocumentTypeItem[];
}

export type ListDocumentsParams = {
  status?: ListDocumentsStatus;
  limit?: number;
  offset?: number;
};

export type ListDocumentsStatus =
  (typeof ListDocumentsStatus)[keyof typeof ListDocumentsStatus];

export const ListDocumentsStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export type ListFindingsParams = {
  severity?: ListFindingsSeverity;
  documentId?: number;
  limit?: number;
};

export type ListFindingsSeverity =
  (typeof ListFindingsSeverity)[keyof typeof ListFindingsSeverity];

export const ListFindingsSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type GetRecentActivityParams = {
  limit?: number;
};