# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProofDesk API - AI-powered contract and document review SaaS
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all documents
 */
export const listDocumentsQueryLimitDefault = 20;
export const listDocumentsQueryOffsetDefault = 0;

export const ListDocumentsQueryParams = zod.object({
  status: zod.enum(["pending", "processing", "completed", "failed"]).optional(),
  limit: zod.coerce.number().default(listDocumentsQueryLimitDefault),
  offset: zod.coerce.number().default(listDocumentsQueryOffsetDefault),
});

export const ListDocumentsResponse = zod.object({
  documents: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      documentType: zod.string(),
      status: zod.enum(["pending", "processing", "completed", "failed"]),
      content: zod.string().nullish(),
      overallRisk: zod.enum(["low", "medium", "high", "critical"]).nullish(),
      summary: zod.string().nullish(),
      findingCount: zod.number(),
      criticalCount: zod.number(),
      createdAt: zod.string(),
      updatedAt: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Upload a new document for review
 */
export const CreateDocumentBody = zod.object({
  title: zod.string(),
  documentType: zod.string(),
  content: zod.string(),
});

/**
 * @summary Get a document by ID
 */
export const GetDocumentParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDocumentResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  documentType: zod.string(),
  status: zod.enum(["pending", "processing", "completed", "failed"]),
  content: zod.string().nullish(),
  overallRisk: zod.enum(["low", "medium", "high", "critical"]).nullish(),
  summary: zod.string().nullish(),
  findingCount: zod.number(),
  criticalCount: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a document
 */
export const DeleteDocumentParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteDocumentResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Trigger AI analysis of a document
 */
export const AnalyzeDocumentParams = zod.object({
  id: zod.coerce.number(),
});

export const AnalyzeDocumentResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  documentType: zod.string(),
  status: zod.enum(["pending", "processing", "completed", "failed"]),
  content: zod.string().nullish(),
  overallRisk: zod.enum(["low", "medium", "high", "critical"]).nullish(),
  summary: zod.string().nullish(),
  findingCount: zod.number(),
  criticalCount: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Get full AI review for a document
 */
export const GetDocumentReviewParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDocumentReviewResponse = zod.object({
  document: zod.object({
    id: zod.number(),
    title: zod.string(),
    documentType: zod.string(),
    status: zod.enum(["pending", "processing", "completed", "failed"]),
    content: zod.string().nullish(),
    overallRisk: zod.enum(["low", "medium", "high", "critical"]).nullish(),
    summary: zod.string().nullish(),
    findingCount: zod.number(),
    criticalCount: zod.number(),
    createdAt: zod.string(),
    updatedAt: zod.string(),
  }),
  findings: zod.array(
    zod.object({
      id: zod.number(),
      documentId: zod.number(),
      documentTitle: zod.string().nullish(),
      category: zod.string(),
      severity: zod.enum(["low", "medium", "high", "critical"]),
      title: zod.string(),
      description: zod.string(),
      recommendation: zod.string(),
      clause: zod.string().nullish(),
      createdAt: zod.string(),
    }),
  ),
  summary: zod.string().nullish(),
  overallRisk: zod.enum(["low", "medium", "high", "critical"]).nullish(),
});

/**
 * @summary List findings across all documents
 */
export const listFindingsQueryLimitDefault = 50;

export const ListFindingsQueryParams = zod.object({
  severity: zod.enum(["low", "medium", "high", "critical"]).optional(),
  documentId: zod.coerce.number().optional(),
  limit: zod.coerce.number().default(listFindingsQueryLimitDefault),
});

export const ListFindingsResponse = zod.object({
  findings: zod.array(
    zod.object({
      id: zod.number(),
      documentId: zod.number(),
      documentTitle: zod.string().nullish(),
      category: zod.string(),
      severity: zod.enum(["low", "medium", "high", "critical"]),
      title: zod.string(),
      description: zod.string(),
      recommendation: zod.string(),
      clause: zod.string().nullish(),
      createdAt: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get dashboard summary statistics
 */
export const GetDashboardSummaryResponse = zod.object({
  totalDocuments: zod.number(),
  completedReviews: zod.number(),
  pendingReviews: zod.number(),
  totalFindings: zod.number(),
  criticalFindings: zod.number(),
  highRiskDocuments: zod.number(),
  avgFindingsPerDoc: zod.number(),
});

/**
 * @summary Get recent document activity feed
 */
export const getRecentActivityQueryLimitDefault = 10;

export const GetRecentActivityQueryParams = zod.object({
  limit: zod.coerce.number().default(getRecentActivityQueryLimitDefault),
});

export const GetRecentActivityResponse = zod.object({
  activities: zod.array(
    zod.object({
      id: zod.number(),
      type: zod.enum(["uploaded", "analyzed", "finding_added"]),
      documentId: zod.number(),
      documentTitle: zod.string(),
      description: zod.string(),
      severity: zod.string().nullish(),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary Get risk level breakdown across all documents
 */
export const GetRiskBreakdownResponse = zod.object({
  bySeverity: zod.array(
    zod.object({
      severity: zod.string(),
      count: zod.number(),
    }),
  ),
  byCategory: zod.array(
    zod.object({
      category: zod.string(),
      count: zod.number(),
    }),
  ),
  byDocumentType: zod.array(
    zod.object({
      documentType: zod.string(),
      count: zod.number(),
    }),
  ),
});