# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProofDesk API - AI-powered contract and document review SaaS
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from "./activityItemType";

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  documentId: number;
  documentTitle: string;
  description: string;
  severity?: string | null;
  createdAt: string;
}