# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProofDesk API - AI-powered contract and document review SaaS
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentOverallRisk } from "./documentOverallRisk";
import type { DocumentStatus } from "./documentStatus";

export interface Document {
  id: number;
  title: string;
  documentType: string;
  status: DocumentStatus;
  content?: string | null;
  overallRisk?: DocumentOverallRisk;
  summary?: string | null;
  findingCount: number;
  criticalCount: number;
  createdAt: string;
  updatedAt: string;
}