# lib/api-zod/src/generated/types/documentList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProofDesk API - AI-powered contract and document review SaaS
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from "./document";

export interface DocumentList {
  documents: Document[];
  total: number;
}