# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProofDesk API - AI-powered contract and document review SaaS
 * OpenAPI spec version: 0.1.0
 */
import type { FindingSeverity } from "./findingSeverity";

export interface Finding {
  id: number;
  documentId: number;
  documentTitle?: string | null;
  category: string;
  severity: FindingSeverity;
  title: string;
  description: string;
  recommendation: string;
  clause?: string | null;
  createdAt: string;
}