# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProofDesk API - AI-powered contract and document review SaaS
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessResponse {
  success: boolean;
}